backend/requirements.txt
asgiref==3.11.1
Django==5.2.13
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
gunicorn==25.3.0
dj-database-url==3.1.2
psycopg2-binary==2.9.12
python-dotenv==1.2.2
pytz==2026.1.post1
sqlparse==0.5.5
fixtures==4.3.2
requests==2.33.1